Moved redislog module to its own sub-crate
