+++ Cargo.toml
@@ -16,20 +16,20 @@ repository = "https://github.com/bolcom/unFTP"
 license = "Apache-2.0"
 readme = "README.md"
 
+[workspace]
+
+[dependencies.redislog]
+path="crates/redislog"
+
 [dependencies]
 async-trait = "0.1.37"
 libunftp = {git = "https://github.com/bolcom/libunftp.git", rev = "52c9a9109a3b62c69ebbc4287db0ca6fdcebbcdf"}
-redis = "0.9.0"
-serde_json = "1.0.57"
-chrono = "0.4.6"
-r2d2 = "0.8.3"
-r2d2_redis = "0.8.0"
 slog = { version = "2.4.1", features = ["max_level_trace", "release_max_level_info"] }
 slog-term = "2.6.0"
 slog-async = "2.3.0"
 hyper = "0.13.7"
 http = "0.2.1"
-prometheus = "0.9.0"
+prometheus = "0.10.0"
 futures = "0.3"
 tokio = { version = "0.2", features = ["full"] }
 clap = "2.33.3"
@@ -56,3 +56,7 @@ static = []
 
 [build-dependencies]
 built = "0.3"
+
+[profile.release]
+debug = false
+# lto = true
