+++ crates/redislog/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "redislog"
+version = "0.1.0"
+authors = ["Hannes de Jager <hdejager@bol.com>"]
+edition = "2018"
+license = "Apache-2.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+chrono = "0.4.6"
+r2d2 = "0.8.3"
+r2d2_redis = "0.8.0"
+redis = "0.9.0"
+serde_json = "1.0.57"
+slog = { version = "2.4.1", features = ["max_level_trace", "release_max_level_info"] }
